Select Procedures from Professional Claim data

select procedure related claims using CPT

procedure_selection(
  std_data_root = wd$std_data_root,
  prof_codes_folder = "prof_clm",
  cpt_map = define_proc_by_cpt,
  test_sas_processed_data_loc = NA,
  n_worker = 2
)

Arguments

std_data_root
professional claim data location

prof_codes_folder
professional claim folder name

cpt_map
should include 3 variable: cpt_cd, e_proc_grp and e_proc_grp_lbl

test_sas_processed_data_loc
this is for testing, only assign value when you want to test, eg. compare the processing results with SAS code Zhaohui Fan created. Otherwise leave it NA. When testing, assign the location of sas pre processed prof_clm data. eg. "/Volumes/George_Surgeon_Projects/medicare_data/sample_npct_std/prof_clm.sas7bdat"

n_worker
number of workers to use in future map

Value

Details

select procedures base on CPT

delete missing NPI procedures(can't link to surgeons)

delete duplicated claims

aggregate cpt_mode (if two cases have same info but diff cpt_mode, aggregate as one case include two cpt_mode code)